refactor(documents): move the feature's IPC layer and types into its folder
The workspace's 19 handlers were inline in the app-wide registrar and its 20
IPC shapes sat in two shared type barrels, so the feature still reached into
four files it did not own. They now live in electron/documents/register.ts and
electron/documents/types.ts, with src/documents/types.ts mirroring the shapes
for the renderer. register.ts sheds 150 lines, shared-types.ts 204, and
main.ts calls one stopAllDocumentWork() instead of three shutdown functions.

validatePath joins the other argument validators in ipc/validate.ts so both
registrars share it, and the `kind !== 'document'` filter that had been copied
into three call sites becomes codeProjects().
