+++ electron/ipc/validate.ts
@@ -1,4 +1,12 @@
 /** Runtime type assertion helpers for IPC handler args. */
+import path from 'path';
+
+/** Reject paths that are non-absolute or attempt directory traversal. */
+export function validatePath(p: unknown, label: string): void {
+  if (typeof p !== 'string') throw new Error(`${label} must be a string`);
+  if (!path.isAbsolute(p)) throw new Error(`${label} must be absolute`);
+  if (p.includes('..')) throw new Error(`${label} must not contain ".."`);
+}
 
 export function assertString(val: unknown, label: string): asserts val is string {
   if (typeof val !== 'string') throw new Error(`${label} must be a string`);
