+++ electron/main.ts
@@ -10,8 +10,7 @@ import { installIpcTracing } from './ipc/trace.js';
 import { startAgentHookRuntime, stopAgentHookRuntime } from './agent-hooks/runtime.js';
 import { killAllAgents } from './ipc/pty.js';
 import { stopAllPlanWatchers } from './ipc/plans.js';
-import { stopAllDocumentRuns, stopAllDocumentWatchers } from './documents/runs.js';
-import { stopAllAsks } from './documents/annotations.js';
+import { stopAllDocumentWork } from './documents/register.js';
 import { stopAllStepsWatchers } from './ipc/steps.js';
 import { verificationRunner } from './ipc/verify.js';
 import { IPC } from './ipc/channels.js';
@@ -290,9 +289,7 @@ app.on('will-quit', () => {
   verificationRunner.cancelAll();
   stopAgentHookRuntime();
   stopAllPlanWatchers();
-  stopAllDocumentWatchers();
-  stopAllDocumentRuns();
-  stopAllAsks();
+  stopAllDocumentWork();
   stopAllStepsWatchers();
 });
