+++ src/documents/types.ts
@@ -1,0 +1,2 @@
+/** The document workspace's IPC shapes, mirrored for the renderer. */
+export type * from '../../electron/documents/types';
