+++ src/ipc/types.ts
@@ -7,23 +7,6 @@ export type {
   CoverageMetricSummary,
   CoverageSummary,
   CreateTaskResult,
-  DocumentAnchor,
-  DocumentAnnotation,
-  DocumentAnnotationEvent,
-  DocumentAnnotationKind,
-  DocumentCandidateRecord,
-  DocumentCandidateSpec,
-  DocumentCandidateStatus,
-  DocumentFileInfo,
-  DocumentFolderInfo,
-  DocumentHistoryEntry,
-  DocumentProjectSetup,
-  DocumentRationale,
-  DocumentRunEvent,
-  DocumentRunRecord,
-  DocumentRunStatus,
-  DocumentScope,
-  DocumentSnapshot,
   EslintQualityFinding,
   EslintQualityResult,
   FileDiffResult,
